refactor(test): convert e2b tests to true integration tests
Remove separation between unit and integration tests. In vm0, vitest
runs as single process with access to real database and services.

Changes:
- Delete separate e2b-service-database.integration.test.ts file
- Refactor e2b-service.test.ts to use REAL internal services
- Only mock external services (E2B SDK, S3 client)
- Use real database, storage service, image service
- Set up proper test data (scopes, agentComposes, agentComposeVersions)
- Use real system template (vm0/claude-code) instead of mocks

Mock Strategy (Only External Services):
- MOCK: E2B SDK (external API), S3 client (Cloudflare R2)
- REAL: Database, Storage Service, Image Service

This is the correct approach for vm0 where tests are integration tests
by nature, not isolated unit tests. We test real service interactions
while isolating external API calls.

All 12 tests passing ✅
